requirements.txt
streamlit == '1.25.0'
pandas == '1.5.0'
numpy == '1.23.4'
ultralytics == '8.0.162'
Pillow == '9.5.0'
opencv-python == '4.8.0.76'
hydralit-components == '1.0.10'


